+++ FBSimulatorControl/Notifiers/FBCoreSimulatorNotifier.m
@@ -1,8 +1,10 @@
 /**
- * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
  *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
  */
 
 #import "FBCoreSimulatorNotifier.h"
@@ -17,7 +19,7 @@
 
 @interface FBCoreSimulatorNotifier ()
 
-@property (nonatomic, readonly, assign) NSUInteger handle;
+@property (nonatomic, readonly, assign) unsigned long long handle;
 @property (nonatomic, readonly, strong) id notifier;
 
 @end
@@ -38,17 +40,16 @@
 
 - (instancetype)initWithNotifier:(id<NSObject, SimDeviceNotifier>)notifier queue:(dispatch_queue_t)queue block:(void (^)(NSDictionary *info))block
 {
-  self = [super init];
-  if (!self) {
-    return nil;
-  }
-
-  _notifier = notifier;
-  _handle = [FBCoreSimulatorNotifier registerNotificationHandler:notifier queue:queue handler:block];
-
-  return self;
+    self = [super init];
+    if (!self) {
+        return nil;
+    }
+    
+    _notifier = notifier;
+    _handle = [FBCoreSimulatorNotifier registerNotificationHandler:notifier queue:queue handler:block];
+    
+    return self;
 }
-
 - (void)terminate
 {
   [self.notifier unregisterNotificationHandler:self.handle error:nil];
@@ -56,15 +57,15 @@
 
 + (NSUInteger)registerNotificationHandler:(id<NSObject, SimDeviceNotifier>)notifier queue:(dispatch_queue_t)queue handler:(void (^)(NSDictionary *))handler
 {
-  if ([notifier respondsToSelector:@selector(registerNotificationHandlerOnQueue:handler:)]) {
-    return [notifier registerNotificationHandlerOnQueue:queue handler:handler];
-  } else {
-    return [notifier registerNotificationHandler:^(NSDictionary *info) {
-      dispatch_async(queue, ^{
-        handler(info);
-      });
-    }];
-  }
+    if ([notifier respondsToSelector:@selector(registerNotificationHandlerOnQueue:handler:)]) {
+        return [notifier registerNotificationHandlerOnQueue:queue handler:handler];
+    } else {
+        return [notifier registerNotificationHandler:^(NSDictionary *info) {
+            dispatch_async(queue, ^{
+                handler(info);
+            });
+        }];
+    }
 }
 
 @end
